fix(credential): not supported api-key way
The API supports two types of credentials, which are OAuth 2.0 and API
keys. These ways are exclusive for each other.

So, if `clientId` is filled in `GoogleApp` whether with or without `apiKey`,
`apiKey` will not be used (API works with OAuth flow).
But, only `apiKey` is filled in `GoogleApp`, OAuth flow will not be worked (API works with public API key).